+++ src/sbbs3/sbbs.h
@@ -391,7 +391,7 @@ public:
 	long	exec_bin(const char *mod, csi_t *csi);
 	void	clearvars(csi_t *bin);
 	void	freevars(csi_t *bin);
-	char**	getstrvar(csi_t *bin, long name);
+	char**	getstrvar(csi_t *bin, int32_t name);
 	int32_t*	getintvar(csi_t *bin, int32_t name);
 	char*	copystrvar(csi_t *csi, char *p, char *str);
 	void	skipto(csi_t *csi, uchar inst);
